requirements.txt
asgiref==3.8.1
beautifulsoup4==4.11.2
dj-database-url==2.0.0
Django==4.2
django-bootstrap4==22.3
django-geojson==4.0.0
django-heroku==0.3.1
django-leaflet==0.28.3
gunicorn==20.1.0
psycopg2==2.9.5
soupsieve==2.4
sqlparse==0.4.3
typing_extensions==4.6.3
tzdata==2022.7
whitenoise==6.4.0
wincertstore==0.2
